refactor: rename request into query
