+++ lib/infrastructure/DatamartDatasource.ts
@@ -2,14 +2,14 @@ import { knexDatamart } from '../common/db/knex-database-connections.js';
 import { logger } from '../common/logger/logger.js';
 import { DatamartQuery } from '../domain/models/DatamartQuery.ts';
 import { DatamartResponse } from '../domain/models/DatamartResponse.ts';
-import { RequestBuilder } from './builder/RequestBuilder.ts';
+import { QueryBuilder } from './builder/QueryBuilder.ts';
 
 export interface DatamartRepository {
   find(_datamartQuery: DatamartQuery): Promise<DatamartResponse>;
 }
 class DatamartRepositoryImpl implements DatamartRepository {
   async find(datamartQuery: DatamartQuery): Promise<DatamartResponse> {
-    const requestBuilder = new RequestBuilder(datamartQuery);
+    const requestBuilder = new QueryBuilder(datamartQuery);
     try {
       // eslint-disable-next-line knex/avoid-injections
       const result = await knexDatamart.raw(requestBuilder.build());
